fix bug: size_t非负的问题

+++ src/observer/sql/operator/insert_physical_operator.cpp
@@ -25,7 +25,7 @@ RC InsertPhysicalOperator::open(Trx *trx)
 {
   RC                  rc;
   std::vector<Record> records(values_list_.size());
-  for (size_t i = 0; i < values_list_.size(); ++i) {
+  for (int i = 0; i < values_list_.size(); ++i) {
     rc = table_->make_record(static_cast<int>(values_list_[i].size()), values_list_[i].data(), records[i]);
     if (rc != RC::SUCCESS) {
       LOG_WARN("failed to make record. rc=%s", strrc(rc));
@@ -33,7 +33,7 @@ RC InsertPhysicalOperator::open(Trx *trx)
     }
   }
 
-  for (size_t i = 0; i < records.size(); ++i) {
+  for (int i = 0; i < records.size(); ++i) {
     rc = trx->insert_record(table_, records[i]);
     if (rc != RC::SUCCESS) {
       LOG_WARN("failed to insert record by transaction. rc=%s", strrc(rc));
