+++ src/observer/sql/operator/update_physical_operator.cpp
@@ -51,7 +51,7 @@ RC UpdatePhysicalOperator::open(Trx *trx)
     // rid 得手动拷贝
     new_record.set_rid(old_record.rid());
     new_record.copy_data(old_record.data(), old_record.len());
-    for (size_t i = 0; i < field_metas_.size(); ++i) {
+    for (int i = 0; i < field_metas_.size(); ++i) {
       new_record.set_field(field_metas_[i].offset(), field_metas_[i].len(), values_[i]);
       if (field_metas_[i].nullable()) {
         auto null_offset = field_metas_[i].offset() + field_metas_[i].len() - 1;
